reformatted as per standard formatter and incorporated review comments.
