+++ spring-boot-modules/spring-boot-hexagonal/src/main/java/com/baeldung/domain/Document.java
@@ -1,21 +1,23 @@
 package com.baeldung.domain;
 
 public class Document {
-	private String id;
-	private String data;
-	
-	public String getId() {
-		return id;
-	}
-	public void setId(String id) {
-		this.id = id;
-	}
-	public String getData() {
-		return data;
-	}
-	public void setData(String data) {
-		this.data = data;
-	}
-	
-	
+    private String id;
+    private String data;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getData() {
+        return data;
+    }
+
+    public void setData(String data) {
+        this.data = data;
+    }
+
 }
