+++ spring-boot-modules/spring-boot-hexagonal/src/main/java/com/baeldung/domain/port/DocumentRepo.java
@@ -1,0 +1,9 @@
+package com.baeldung.domain.port;
+
+import com.baeldung.domain.Document;
+
+public interface DocumentRepo {
+    void storeDocument(Document document);
+
+    Document findDocumentById(String id);
+}
