+++ spring-boot-modules/spring-boot-hexagonal/src/main/java/com/baeldung/domain/port/DocumentService.java
@@ -1,0 +1,11 @@
+package com.baeldung.domain.port;
+
+import com.baeldung.domain.Document;
+
+public interface DocumentService {
+
+    void createDocument(Document document);
+
+    Document findById(String id);
+
+}
