--- spring-boot-modules/spring-boot-hexagonal/src/main/java/com/baeldung/port/DocumentRepo.java
@@ -1,10 +1,0 @@
-package com.baeldung.port;
-
-import com.baeldung.domain.Document;
-
-public interface DocumentRepo {
-	void storeDocument(Document document);
-	
-	Document findDocumentById(String id);
-}
-
