--- spring-boot-modules/spring-boot-hexagonal/src/main/java/com/baeldung/port/DocumentService.java
@@ -1,11 +1,0 @@
-package com.baeldung.port;
-
-import com.baeldung.domain.Document;
-
-public interface DocumentService {
-	
-	void createDocument(Document document);
-	
-	Document findById(String id);
-	
-}
